Reorganize code
Create a file for patching each different version of ActiveRecord.
